requirements.txt
azure-functions
azure-storage-blob==12.19.0
requests>=2.31.0
urllib3>=2.0.0
